Seperate  library and encoder into seperate files
